M3-203 Refactor : 지도 위젯 폴더 구성 변경
